+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 1.0.2 (not yet released)
 
+- Add "positiveInteger" and "arrayOfPositiveInteger" option types that only
+  accept positive integers.
+
 - Add "integer" and "arrayOfInteger" option types that accepts only integers.
   Note that, for better or worse, these do NOT accept: "0x42" (hex), "1e2"
   (with exponent) or "1.", "3.0" (floats).